COMMERCE-919 Refactor packages (no logic change)
