+++ commerce-openapi-core-impl/src/test/java/com/liferay/commerce/openapi/core/internal/jaxrs/nested/dto/ProductDTO.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.commerce.openapi.core.internal.jaxrs.nested.util;
+package com.liferay.commerce.openapi.core.internal.jaxrs.nested.dto;
 
 import java.util.Collection;
 import java.util.Collections;
